// resources/js/api/generated/models/getApiV1ServersServeridStacksStackname200ServicesItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1ServersServeridStacksStackname200ServicesItemContainersItem } from './getApiV1ServersServeridStacksStackname200ServicesItemContainersItem';

export type GetApiV1ServersServeridStacksStackname200ServicesItem = {
  containers: GetApiV1ServersServeridStacksStackname200ServicesItemContainersItem[];
  image?: string;
  name: string;
  ports?: string[];
};